+++ provider/caleta/config.go
@@ -14,6 +14,16 @@ type AuthConf struct {
 	// VerificationKey Used to verify incoming requests
 	VerificationKey string `mapstructure:"verification_key"`
 }
+type GameLaunchType string
+
+const (
+	Static  GameLaunchType = "static"
+	Request GameLaunchType = "request"
+)
+
+type caletaConf struct {
+	GameLaunchType GameLaunchType `mapstructure:"game_launch_type"`
+}
 
 // getAuthConf parse provider specific auth configuration
 func getAuthConf(c configs.ProviderConf) (AuthConf, error) {
@@ -23,4 +33,25 @@ func getAuthConf(c configs.ProviderConf) (AuthConf, error) {
 		return auth, err
 	}
 	return auth, nil
+}
+
+// getCaletaConf parse provide specific configuration
+func getCaletaConf(c configs.ProviderConf) (caletaConf, error) {
+	var cc caletaConf
+	if c.ProviderSpecific != nil {
+		err := mapstructure.Decode(c.ProviderSpecific, &cc)
+		if err != nil {
+			return cc, err
+		}
+		if cc.GameLaunchType != Static && cc.GameLaunchType != Request {
+			// Default to Static
+			cc.GameLaunchType = Static
+		}
+	} else {
+		// Default to Static
+		cc = caletaConf{
+			GameLaunchType: Static,
+		}
+	}
+	return cc, nil
 }